requirements.txt
flask==3.0.2
playwright==1.48.0
python-dateutil==2.9.0.post0
requests>=2.32.5

